+++ src/MarginTrading.Backend.Services/Extensions/DependencyInjectionExtensions.cs
@@ -7,7 +7,7 @@ using System;
 using System.Diagnostics.CodeAnalysis;
 using Autofac;
 
-using MarginTrading.Backend.Services.Snapshot;
+using MarginTrading.Backend.Services.Snapshots;
 using MarginTrading.Common.Extensions;
 
 namespace MarginTrading.Backend.Services.Extensions
