+++ src/MarginTrading.Backend.Services/Services/OrdersProvider.cs
@@ -10,7 +10,7 @@ using MarginTrading.Backend.Core.Orders;
 using MarginTrading.Backend.Core.Services;
 using MarginTrading.Backend.Core.Trading;
 using MarginTrading.Backend.Services.Extensions;
-using MarginTrading.Backend.Services.Snapshot;
+using MarginTrading.Backend.Services.Snapshots;
 
 namespace MarginTrading.Backend.Services.Services
 {
