+++ src/MarginTrading.Backend.Services/Services/PositionsProvider.cs
@@ -9,7 +9,7 @@ using Common.Log;
 using MarginTrading.Backend.Core.Orders;
 using MarginTrading.Backend.Core.Services;
 using MarginTrading.Backend.Services.Extensions;
-using MarginTrading.Backend.Services.Snapshot;
+using MarginTrading.Backend.Services.Snapshots;
 
 namespace MarginTrading.Backend.Services.Services
 {
