+++ src/MarginTrading.Backend.Services/Snapshots/AsSoonAsPossibleStrategy.cs
@@ -2,7 +2,7 @@ using System.Threading.Tasks;
 
 using MarginTrading.Backend.Core.Snapshots;
 
-namespace MarginTrading.Backend.Services.Snapshot;
+namespace MarginTrading.Backend.Services.Snapshots;
 
 public class AsSoonAsPossibleStrategy(IEnvironmentValidator environmentValidator) : IEnvironmentValidationStrategy
 {
