+++ src/MarginTrading.Backend.Services/Snapshots/IFakeSnapshotService.cs
@@ -10,7 +10,7 @@ using MarginTrading.Backend.Contracts.Orders;
 using MarginTrading.Backend.Contracts.Positions;
 using MarginTrading.Backend.Contracts.Snow.Prices;
 
-namespace MarginTrading.Backend.Services.Snapshot
+namespace MarginTrading.Backend.Services.Snapshots
 {
     public interface IFakeSnapshotService
     {
