+++ src/MarginTrading.Backend.Services/Snapshots/LoggingTradingEngineRawSnapshotsRepository.cs
@@ -8,7 +8,7 @@ using MarginTrading.Backend.Core.Snapshots;
 
 using Microsoft.Extensions.Logging;
 
-namespace MarginTrading.Backend.Services.Snapshot;
+namespace MarginTrading.Backend.Services.Snapshots;
 
 public class LoggingTradingEngineRawSnapshotsRepository(
     ITradingEngineRawSnapshotsRepository decoratee,
