+++ src/MarginTrading.Backend.Services/Snapshots/TradingEngineRawSnapshotsAdapter.cs
@@ -8,7 +8,7 @@ using MarginTrading.Backend.Core.Services;
 using MarginTrading.Backend.Core.Snapshots;
 using MarginTrading.Backend.Services.Mappers;
 
-namespace MarginTrading.Backend.Services.Snapshot;
+namespace MarginTrading.Backend.Services.Snapshots;
 
 public class TradingEngineRawSnapshotsRepository(
     ITradingEngineSnapshotsRepository tradingEngineSnapshotsRepository) : ITradingEngineRawSnapshotsRepository
