+++ tests/MarginTradingTests/AssetDayOffTests.cs
@@ -15,7 +15,7 @@ using MarginTrading.Backend.Services.Infrastructure;
 using MarginTrading.Common.Services;
 using MarginTrading.AssetService.Contracts;
 using MarginTrading.AssetService.Contracts.Scheduling;
-using MarginTrading.Backend.Services.Snapshot;
+using MarginTrading.Backend.Services.Snapshots;
 
 using Microsoft.FeatureManagement;
 using Moq;
