+++ tests/MarginTradingTests/DraftSnapshotKeeperTests.cs
@@ -14,7 +14,7 @@ using MarginTrading.Backend.Core.Repositories;
 using MarginTrading.Backend.Core.Snapshots;
 using MarginTrading.Backend.Core.Trading;
 using MarginTrading.Backend.Services;
-using MarginTrading.Backend.Services.Snapshot;
+using MarginTrading.Backend.Services.Snapshots;
 
 using MarginTradingTests.Helpers;
 using Moq;
