+++ tests/MarginTradingTests/Infrastructure/SnapshotValidationServiceTests.cs
@@ -15,7 +15,7 @@ using MarginTrading.Backend.Core.Snapshots;
 using MarginTrading.Backend.Core.Trading;
 using MarginTrading.Backend.Services;
 using MarginTrading.Backend.Services.Infrastructure;
-using MarginTrading.Backend.Services.Snapshot;
+using MarginTrading.Backend.Services.Snapshots;
 
 using Moq;
 using NUnit.Framework;
