+++ tests/MarginTradingTests/PositionsProviderTests.cs
@@ -9,7 +9,7 @@ using Autofac.Extras.Moq;
 using MarginTrading.Backend.Core.Orders;
 using MarginTrading.Backend.Services;
 using MarginTrading.Backend.Services.Services;
-using MarginTrading.Backend.Services.Snapshot;
+using MarginTrading.Backend.Services.Snapshots;
 
 using Moq;
 using NUnit.Framework;
